+++ src/Core/DataFactory/State/dataFactoryState.ts
@@ -1,0 +1,9 @@
+import { Factory } from '../Types/factory';
+
+export interface DataFactoryState {
+    isActiveFactory?: Factory;
+    factoryScope?: Record<string, unknown>;
+    factories: {
+        [key: string]: Factory;
+    };
+}
