+++ src/Core/DataFactory/index.ts
@@ -1,1 +1,5 @@
-export * from './dataFactory';
+export * from './Types/factory';
+export * from './Hooks/useFactory';
+export * from './Functions/createDataFactory';
+export * from './Components/AddMenu';
+export * from './Components/FactoryComponent';
